chore(ts-tests): refactor `ts-tests/*` to follow pnpm workspace design
