+++ .github/workflows/ci.yml
@@ -189,16 +189,3 @@ jobs:
       - uses: codecov/codecov-action@81cd2dc8148241f03f5839d295e000b8f761e378 # tag=v3.1.0
         with:
           fail_ci_if_error: true
-
-  test_pnpm_e2e: # TODO(fuxingloh): move to PNPM
-    name: Test (pnpm e2e)
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@2541b1294d2704b0964813337f33b291d3f8596b # tag=v3.0.2
-
-      - uses: actions/setup-node@2fddd8803e2f5c9604345a0b591c3020ee971a93 # tag=v3.4.1
-        with:
-          node-version: 16
-          cache: 'npm'
-          cache-dependency-path: 'ts-tests/package-lock.json'
-      - run: cd ts-tests && npm ci && npm t
