--- ts-tests/.prettierrc
@@ -1,6 +1,0 @@
-{
-  "semi": true,
-  "singleQuote": true,
-  "trailingComma": "none",
-  "printWidth": 120
-}
