--- ts-tests/jest.config.ts
@@ -1,11 +1,0 @@
-export default {
-  preset: 'ts-jest',
-  verbose: true,
-  clearMocks: true,
-  testTimeout: 180000,
-  transform: {
-    '^.+\\.ts?$': 'ts-jest',
-    '^.+\\.js?$': 'babel-jest'
-  },
-  transformIgnorePatterns: ['node_modules/?!(@polkadot)']
-};
