+++ ts-tests/package.json
@@ -2,29 +2,19 @@
   "name": "ts-tests",
   "version": "0.0.0",
   "scripts": {
-    "test": "hardhat compile && jest"
+    "compile": "hardhat compile",
+    "test": "jest"
+  },
+  "jest": {
+    "displayName": "test",
+    "preset": "@birthdayresearch/sticky-turbo-jest"
   },
   "devDependencies": {
-    "@babel/core": "^7.18.6",
-    "@babel/preset-env": "^7.18.6",
+    "@birthdayresearch/sticky-turbo-jest": "0.2.0",
+    "@defimetachain/typescript": "workspace:*",
     "@polkadot/api": "^8.8.2",
-    "@types/jest": "^27.5.2",
-    "babel-jest": "^27.5.1",
-    "cross-fetch": "^3.1.5",
-    "eslint": "^7.32.0",
-    "eslint-config-standard-jsx": "^10.0.0",
-    "eslint-config-standard-with-typescript": "^21.0.1",
     "ethers": "^5.4.6",
     "hardhat": "^2.11.1",
-    "jest": "^27.5.1",
-    "lint-staged": "^13.0.3",
-    "nock": "^13.2.8",
-    "shuffle-seed": "^1.1.6",
-    "solc": "^0.8.3",
-    "testcontainers": "^8.12.0",
-    "ts-jest": "^27.1.5",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.7.4",
-    "wait-for-expect": "^3.0.2"
+    "testcontainers": "^8.12.0"
   }
 }
