+++ ts-tests/tsconfig.json
@@ -1,21 +1,7 @@
 {
+  "extends": "@defimetachain/typescript/tsconfig.json",
   "compilerOptions": {
-    "lib": ["es2020"],
-    "target": "es2020",
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "strict": true,
-    "esModuleInterop": true,
     "baseUrl": "./",
-    "experimentalDecorators": true,
-    "incremental": true,
-    "allowSyntheticDefaultImports": true,
-    "emitDecoratorMetadata": true,
-    "resolveJsonModule": true,
-    "forceConsistentCasingInFileNames": true,
-    "skipLibCheck": true,
-    "importHelpers": true,
-    "sourceMap": true
-  },
-  "exclude": ["node_modules/"]
+    "rootDir": "./"
+  }
 }
